Get tabs for each user

+++ frontend/src/components/Home/Home.tsx
@@ -1,38 +1,51 @@
 import { useState, useEffect, useRef } from 'react';
 import { AlphaTabApi, Settings } from '@coderline/alphatab';
+import useUser from '../App/useUser';
 import './Home.css';
 
-import ElectricRed from '../../testTab/ElectricRed.gp5';
-import Altitudes from '../../testTab/Altitudes.gp5';
-import InDeathIsDeath from '../../testTab/in_death_is_death.gp5';
-import whirlTHIS from '../../testTab/whirlTHIS.gp5';
+// import ElectricRed from '../../testTab/electric_red.gp5';
+// import Altitudes from '../../testTab/Altitudes.gp5';
+// import InDeathIsDeath from '../../testTab/in_death_is_death.gp5';
+// import whirlTHIS from '../../testTab/whirlTHIS.gp5';
 
 
 export default function Home() {
     const elementRef = useRef<HTMLDivElement>(null);
     const [api, setApi] = useState<AlphaTabApi>();
+    const { user } = useUser();
+    const [tabFilePath, setTabFilePath] = useState<string | null>(null);
 
     useEffect(() => {
-        const api = new AlphaTabApi(elementRef.current!, {
-        core: {
-            file: whirlTHIS,
-            fontDirectory: '/font/'
-        },
-        player: {
-            enablePlayer: true,
-            enableCursor: true,
-            enableUserInteraction: true,
-            soundFont: '/soundfont/sonivox.sf2'
+        if (user && user.tabs && user.tabs.length > 0) {
+            const fileURL = user.tabs[0].fileURL.replace('../', '/');
+            console.log(`${fileURL}`)
+            setTabFilePath(fileURL);
         }
-        } as Settings);
+    }, [user]);
 
-        setApi( api );
+    useEffect(() => {
+        if (tabFilePath) {
+            const api = new AlphaTabApi(elementRef.current!, {
+                core: {
+                    file: tabFilePath,
+                    fontDirectory: '/font/'
+                },
+                player: {
+                    enablePlayer: true,
+                    enableCursor: true,
+                    enableUserInteraction: true,
+                    soundFont: '/soundfont/sonivox.sf2'
+                }
+                } as Settings);
 
-        return () => {
-            console.log('destroy', elementRef)
-            api.destroy();
+                setApi( api );
+
+                return () => {
+                    console.log('destroy', elementRef)
+                    api.destroy();
+                }
         }
-    }, []);
+    }, [tabFilePath]);
 
     function playPause() {
         api?.playPause();
